requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
black==24.8.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==42.0.8
daphne==4.1.2
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
langdetect==1.0.9
multidict==6.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.35.12
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
playsound==1.3.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stripe==10.6.0
tomli==2.0.1
tqdm==4.66.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.11.1
zope.interface==7.0.3
